Updated headers for Worker and WorkQueue.
Scheduler in progress. Every Worker thread is going to have a
real time thread priority (value is still to be determined) which
makes the whole system adaptive and scalable. Communication
is designed ot be non-blocking and thus every single connection
can be run in independently with the help of libevent2.